@@ -1,8 +1,7 @@
-from __future__ import absolute_import
-from __future__ import unicode_literals
-
 import argparse
-import io
+from typing import Optional
+from typing import Sequence
+from typing import Type
 
 import mistune
 import pygments.formatters
@@ -11,10 +10,10 @@ import pygments.util
 
 
 class CodeRenderer(mistune.Renderer):
-    def block_code(self, code, lang):
+    def block_code(self, code: str, lang: str) -> str:
         try:
             lexer = pygments.lexers.get_lexer_by_name(lang, stripnl=False)
-            cssclass = 'highlight {}'.format(lang)
+            cssclass = f'highlight {lang}'
         except pygments.util.ClassNotFound:
             lexer = pygments.lexers.get_lexer_by_name('text', stripnl=False)
             cssclass = 'highlight'
@@ -22,18 +21,22 @@ class CodeRenderer(mistune.Renderer):
         return pygments.highlight(code, lexer=lexer, formatter=formatter)
 
 
-def highlight(doc, Renderer=CodeRenderer):
+def highlight(
+        doc: str,
+        Renderer: Type[mistune.Renderer] = CodeRenderer,
+) -> str:
     return mistune.Markdown(Renderer(escape=True, hard_wrap=False))(doc)
 
 
-def main(argv=None):
+def main(argv: Optional[Sequence[str]] = None) -> int:
     parser = argparse.ArgumentParser()
     parser.add_argument('filename', default='/dev/stdin')
     args = parser.parse_args(argv)
 
-    with io.open(args.filename) as f:
+    with open(args.filename) as f:
         hl = highlight(f.read())
-        print('<!doctype html><html><body>{}</body></html>'.format(hl))
+        print(f'<!doctype html><html><body>{hl}</body></html>')
+    return 0
 
 
 if __name__ == '__main__':
